requirements.txt
numpy
pandas
matplotlib
pyyaml
scikit-learn
gstools>=1.3
pykrige>=1.6